+++ influence_benchmark/RL/KTO.py
@@ -1,4 +1,3 @@
-# TODO rename this to multi-step KTO
 import json
 
 from influence_benchmark.RL.base_iteration import BaseIteration
